Changed float parameters back to double parameters
While debugging to try to determine the cause of the random and
inaccurate # for reconstruction_histories, many of the parameters for
the MLP endpoints and DROP host/GPU functions were changed from double
to float.  In addition, some of the DROP functions were always using
float parameters even though many of the calculations they were used in
included other terms that were all or partially double parameters.  All
of these float parameters have now been changed to double parameters,
the only float terms that remain are the various host/GPU vectors/arrays
containing the proton data information.  In this case it is probably
necessary to keep these as float as these will consume half the global
GPU memory than double arrays would.  However, there is considerably
more global GPU memory on the cluster GPUs so these float variables may
eventually be converted to double arrays, but since the data read from
disk are all float values, there may be no benefit to doing this.  In
this case, it would not make sense to waste the additional memory needed
to store the double arrays and these should remain float.  A detailed
analysis of the data consumption per history for each part of the
program needs to be performed to determine the memory consumption at
each stage of reconstruction and this will give the program properties
needed to make this and similarly related decisions about parameter
types and data transfer methods.