Added new GPU image reconstruction parameters to config file/read functions
The various parameters we want to change in testing the GPU image
reconstruction accuracy/performance have been added to Configurations.h
to bring the new version of the program closer to being ready for future
use.  Some additional cosmetic changes were made as well.

This commit also includes additions to the _Blake code being used to run
GPU image reconstruction implementations.  This program now has the
automated creation/naming of unique output directories which are named
with the execution date and "_i" appended to the folder name for
increasing values of i until there is no existing folder with that name.
This makes running automated tests much easier.