Verified automatic permissions, unique folder naming, and execution results writing routines
The various functions that implement automatically setting permissions
of the newly reconstructed data/images, creating a unique folder with
the execution date in the name followed by "_n" for increasing n until
the folder does not exist, and execution results are written to a single
.csv for all results and a .txt file specific to each reconstruction and
written along with the reconstructed images.

I addition, the data sets (organized_data) and reconstructed data/images
(reconstructed_data) are now organized with the exact same structure
except slightly varying path to where organized_data/reconstructed_data
are both located.  The subdirectories under these 2 directories are then
identical making it easier to transfer data from one node/workstation to
the next since the cluster nodes do not have interactive displays in
which the reconstructed images can be viewed, so this data must always
be transferred through the network from <compute node> -> WHartnell ->
Kodiak-> the workstation or remote computer from which Kodiak is being
accessed.