Reconstruction working with experimental Edge Phantom data
The issue preventing properly executing GPU image reconstruction using
the large experimental Edge Phantom data set appears to have been
resolved by increasing the MAX_INTERSECTIONS for an MLP path to 1280.
This was initially increased to 1024 and this ran properly for several
hours, but this eventually stopped working.  It is unclear at the moment
why this occurred, there were many things changed in this time including
the height of the reconstruction volume and 1 aspect of the MLP
endpoints routine, but restoring these to their previous implementations
did not appear to help.  The radius of the reconstruction volume appears
to have the most impact on this issue and this was increased during
testing so it is possible that this or 1 of the other parameters had in
fact been changed between the time when it worked and when it ceased to
work, only time and more investigation will tell.

When it was running properly there were at least several intervals
between executions and it appears that error tend to occur in groups of
executions particularly when they are ran with little time between.  I
am beginning to wonder if the repeated consecutive executions are
overheating the GPU and causing malfunctions that would not normally
occur because there does seem to be some correlation between these
events which I had been viewing as random, but perhaps they are not.