Adding TVS implementation, basic structure only
Began adding the basic structure for the TVS algorithm, using host code
approach with for loops but placed inside a __device__function since
most of it can be done in parallel on a per voxel basis.  Hence, the
current __device__function is not valid and would not work, but this
provides the basis for continuing work which will iteratively replace
the for loops with thread dependent actions.