First draft implementation of host & GPU TVS
Implemented TVS using both host and GPU approaches, both of which are
applied before each DROP update.  The GPU implementation turns out to be
more difficult than expected due to the few sequential portions, which
are difficult to do without transferring image arrays to/from the GPU.

Some of the implementations clearly will not work and will be deleted,
but this commit includes these so these can be recovered in the future
if an idea to make them feasible arises.