Cleaned up TVS and setup program for TVS testing
Removed the TVS functions that clearly will not work and finalized the
draft implementation of TVS with 4 total versions consisting of TVS
applied before or after DROP updates are applied and a host only and
parallel version of each of these.  There are 3 boolean variables
controlling TVS, 1 to turn it on/off, 1 to specify the order of
TVS/DROP, and 1 to control if the host only or parallel implementation
is used.  Testing of these can begin once the logic has been verified
and the data management and host/GPU transfer/storage approach has been
tested.

This is almost entirely related to the parallel version which is
avoiding ALL intermediate host/GPU transfers of perturbation vectors and
perturbed images for different values of beta by exploiting a
mathematical pattern arising due to the halving of beta and using a
single thread on the GPU to perform the sequential portions of the code
by storing their return values in a single element array, transferring
these to/from GPU, and using this value to control subsequent kernel
launches instead of having to transfer the image to the host, calculate
these values, and apply the new value of beta and transfer the resulting
image back to the GPU.