Modified foundation for recon history ordering and added block ordering scheme
(1) This commit includes routines to calculate the # of blocks used in
the block iterative reconstruction algorithms based on the block size
parameter and # of histories entering reconstruction and use this to
precalculate the # of histories in each block and the index
corresponding to the 1st history in each block.  By precalculating and
storing this information in vectors, not only can the current approach
in which this information is calculated for each block and repeated for
each iteration, but the order in which the blocks are used in
reconstruction can also easily be varied.  The schemes for optionally
reordering blocks made available in this commit are rotation left/right
and randomized shuffling.

(2) This commit also includes the preliminary work to automate the
entirety of the reconstruction history ordering scheme once the AKS
prime # testing algorithm has been incorporated into the program.  The
history ordering scheme implemented before the MLP + image
reconstruction routines were parallelized reordered histories such that
successive proton histories had paths with roughly orthogonal projection
angles.  This exploited the fact that the projection data files have
roughly the same # of histories and are read in order of increasing
projection angle, so prior to reconstruction the histories are roughly
in order of increasing projection angle and orthogonal protons have an
index separation of roughly N/4, where N is the total # of histories
entering reconstruction.  To order the histories for reconstruction such
that successive protons are roughly orthogonal and each history appears
once and only once, the index of successive protons from the original
ordering were calculated as

i(n+1)=i(n) + p mod(N)

where p is chosen to be the 1st prime # p>N/4.  However, without an
efficient implementation of a prime # test, p was determined using the
prime # tests available on the web and, therefore, this scheme was only
tested in a handful of cases but not used in general.  I have now found
implementations of the AKS prime testing algorithm, which is one of the
only deterministic polynomial time algorithms for this test.  I am
currently testing these implementations separately to determine the most
efficient and decide if the required computation time for such a test is
feasible and this commit contains the modifications to the original
implementation so the history ordering is ready to use once the prime
test has been added.