Incorporated exact projection angle data from continuous scans and FBP improvements
Continuous scan data allows exact projection angle to be approximated
using the gantry angle rotation speed and each history's time stamp.
Robert's version of preprocessing appends these float values to the end
of each file and the "magic number" is "PTCN" instead of "PCTD" when
this data is present.  The data reads then check this and a Boolean
specifies if this data should be used when present.  When it is present
and used, the recon_volume_intersections_GPU kernel is passed the float
array actual_projection_angles_d replacing the integer array
gantry_angle_d but they are used in exactly the same way so the code
remains the same, only the array type is changed.  The recon volume
intersections are then more accurately calculated for continuous data
than if they only used discrete angles.

Since the only change is the array type, the 2 kernels can be combined
into a single template kernel and this will likely appear in the next
commit.

FBP was also looked at again after finding that the original
implementation had used a different coordinate system and inverted z
since it had coordinate/voxel conversions hardcoded instead of calling
the standardized functions for these tasks.  After replacing hard coded
calculations with functions, the bin/voxel identification and
backpojection summations were modified to mirror that of the old
reconstruction program, primarily to test if this was in any way
responsible for the better FBP quality from the old program.  Some of
the utv/xyz and bin/voxel conversions are still hard coded and will be
replaced by standardized functions as well in a future commit.