Removing Notepad++ backup folder/files
Notepad++ automatically creates an "nppBackup" directory and saves a
backup of a file modified/saved when this is done in Notepad++.  This
folder and the corresponding backup files have now been added to
.gitignore so these will not appear in future commits, but the
folder/files were included in previous commits and appear in the
repository.  This commit is intended to remove these from the repository
since although these can be very helpful for recovery and error related
issues, Notepad users only need a local copy of these files on their
personal computer and it is wholly unnecessary to share these with other
users.